Move package parsing to package.ts
