+++ src/components/manager.ts
@@ -4,7 +4,6 @@ import * as path from 'path'
 import * as fs from 'fs'
 import * as tmp from 'tmp'
 import * as chokidar from 'chokidar'
-import {latexParser} from 'latex-utensils'
 import * as utils from '../utils/utils'
 // import {InputFileRegExp} from '../utils/inputfilepath'
 
@@ -410,76 +409,6 @@ export class Manager {
 
     watchPdfFile(pdfFileUri: vscode.Uri) {
         this.pdfWatcher.watchPdfFile(pdfFileUri)
-    }
-
-    /**
-     * Updates the Manager cache for packages used in `file` with `nodes`.
-     * If `nodes` is `undefined`, `content` is parsed with regular expressions,
-     * and the result is used to update the cache.
-     *
-     * @param file The path of a LaTeX file.
-     * @param nodes AST of a LaTeX file.
-     * @param content The content of a LaTeX file.
-     */
-    updateUsepackage(file: string, nodes?: latexParser.Node[], content?: string) {
-        if (nodes !== undefined) {
-            this.updateUsepackageNodes(file, nodes)
-        } else if (content !== undefined) {
-            const pkgReg = /\\usepackage(\[[^[\]{}]*\])?{(.*)}/gs
-
-            while (true) {
-                const result = pkgReg.exec(content)
-                if (result === null) {
-                    break
-                }
-                const packages = result[2].split(',').map(packageName => packageName.trim())
-                const options = (result[1] || '[]').slice(1,-1).replaceAll(/\s*=\s*/g,'=').split(',').map(option => option.trim())
-                const optionsNoTrue = options.filter(option => option.includes('=true')).map(option => option.replace('=true', ''))
-                packages.forEach(packageName => this.pushUsepackage(file, packageName, [...options, ...optionsNoTrue]))
-            }
-        }
-    }
-
-    private updateUsepackageNodes(file: string, nodes: latexParser.Node[]) {
-        nodes.forEach(node => {
-            if ( latexParser.isCommand(node) && (node.name === 'usepackage' || node.name === 'documentclass') ) {
-                let options: string[] = []
-                node.args.forEach(arg => {
-                    if (latexParser.isOptionalArg(arg)) {
-                        options = arg.content.filter(latexParser.isTextString).filter(str => str.content !== ',').map(str => str.content)
-                        const optionsNoTrue = options.filter(option => option.includes('=true')).map(option => option.replace('=true', ''))
-                        options = [...options, ...optionsNoTrue]
-                        return
-                    }
-                    for (const c of arg.content) {
-                        if (!latexParser.isTextString(c)) {
-                            continue
-                        }
-                        c.content.split(',').forEach(packageName => this.pushUsepackage(file, packageName, options, node))
-                    }
-                })
-            } else {
-                if (latexParser.hasContentArray(node)) {
-                    this.updateUsepackageNodes(file, node.content)
-                }
-            }
-        })
-    }
-
-    private pushUsepackage(fileName: string, packageName: string, options: string[], node?: latexParser.Command) {
-        packageName = packageName.trim()
-        if (packageName === '') {
-            return
-        }
-        if (node?.name === 'documentclass') {
-            packageName = 'class-' + packageName
-        }
-        const cache = this.extension.cacher.getCachedContent(fileName)
-        if (cache === undefined) {
-            return
-        }
-        cache.elements.package = cache.elements.package || {}
-        cache.elements.package[packageName] = options
     }
 
     private registerSetEnvVar() {
