+++ src/providers/completer/command.ts
@@ -211,7 +211,7 @@ export class Command implements IProvider {
      */
     update(file: string, nodes?: latexParser.Node[], content?: string) {
         // First, we must update the package list
-        this.extension.manager.updateUsepackage(file, nodes, content)
+        this.extension.completer.package.updateUsepackage(file, nodes, content)
         // Remove newcommand cmds, because they will be re-insert in the next step
         this.definedCmds.forEach((entry,cmd) => {
             if (entry.file === file) {
